+++ ANF Code Test/Models/ContentItem.swift
