+++ ANF Code Test/Models/PromoItem.swift
